service files for the pages are created
