+++ app/backend/src/main/java/com/app/gamereview/controller/ImageController.java
@@ -1,5 +1,5 @@
 package com.app.gamereview.controller;
-
+/* 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.FileSystemResource;
 import org.springframework.http.MediaType;
@@ -42,3 +42,4 @@ public class ImageController {
 	}
 
 }
+*/
